+++ nym-wallet/src/requests/queries.ts
@@ -9,7 +9,7 @@ import {
   WrappedDelegationEvent,
   PendingIntervalEvent,
 } from '@nymproject/types';
-import { Interval, TNodeDescription } from 'src/types';
+import { Interval, TGatewayReport, TNodeDescription } from 'src/types';
 import { invokeWrapper } from './wrapper';
 
 export const getAllPendingDelegations = async () =>
@@ -48,3 +48,6 @@ export const getNodeDescription = async (host: string, port: number) =>
 
 export const getPendingIntervalEvents = async () =>
   invokeWrapper<PendingIntervalEvent[]>('get_pending_interval_events');
+
+export const getGatewayReport = async (identity: string) =>
+  invokeWrapper<TGatewayReport>('gateway_report', { identity });