+++ service-providers/network-statistics/src/main.rs
@@ -1,9 +1,9 @@
 // Copyright 2022 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: Apache-2.0
 
-use std::path::PathBuf;
-
 use api::NetworkStatisticsAPI;
+use logging::setup_logging;
+use std::path::PathBuf;
 
 mod api;
 mod storage;
@@ -21,24 +21,6 @@ async fn main() {
         .await
         .expect("Could not ignite stats api service");
     api.run().await;
-}
-
-fn setup_logging() {
-    let mut log_builder = pretty_env_logger::formatted_timed_builder();
-    if let Ok(s) = ::std::env::var("RUST_LOG") {
-        log_builder.parse_filters(&s);
-    } else {
-        // default to 'Info'
-        log_builder.filter(None, log::LevelFilter::Info);
-    }
-
-    log_builder
-        .filter_module("hyper", log::LevelFilter::Warn)
-        .filter_module("tokio_reactor", log::LevelFilter::Warn)
-        .filter_module("reqwest", log::LevelFilter::Warn)
-        .filter_module("mio", log::LevelFilter::Warn)
-        .filter_module("want", log::LevelFilter::Warn)
-        .init();
 }
 
 /// Returns the default base directory for the storefile.