+++ validator-api/src/epoch_operations/mod.rs
@@ -30,6 +30,7 @@ pub(crate) mod error;
 mod helpers;
 
 use crate::epoch_operations::helpers::stake_to_f64;
+use crate::node_status_api::ONE_DAY;
 use error::RewardingError;
 use mixnet_contract_common::mixnode::MixNodeDetails;
 use task::ShutdownListener;
@@ -49,9 +50,6 @@ impl From<MixnodeToReward> for ExecuteMsg {
         }
     }
 }
-
-// // Epoch has all the same semantics as interval, but has a lower set duration
-// type Epoch = Interval;
 
 pub struct RewardedSetUpdater {
     nymd_client: Client<SigningNymdClient>,
@@ -268,8 +266,8 @@ impl RewardedSetUpdater {
             log::info!("Advanced the epoch and updated the rewarded set... SUCCESS");
         }
 
-        log::info!("Puring all node statuses from the storage...");
-        let cutoff = (epoch_end - Duration::from_secs(86400)).unix_timestamp();
+        log::info!("Purging old node statuses from the storage...");
+        let cutoff = (epoch_end - 2 * ONE_DAY).unix_timestamp();
         self.storage.purge_old_statuses(cutoff).await?;
 
         Ok(())